% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIRberedHjelpefunk.R
\name{abonnementBeredsk}
\alias{abonnementBeredsk}
\title{Funksjon som produserer rapporten som skal sendes til mottager.
Nødvendig info til Covid-rapporten er reshID og enhetsNivaa}
\usage{
abonnementBeredsk(
  rnwFil,
  brukernavn = "beredskap",
  valgtRHF = "tom",
  reshID = 0,
  enhetsNivaa = "Alle",
  nivaaNavn = "tom",
  Rpakke = "intensivberedskap"
)
}
\arguments{
\item{rnwFil}{Navn på fila som skal kjøres. Angis MED filending (\emph{dvs "filnavn.Rnw"})}

\item{valgtRHF}{Fases ut, men beholdes for at gamle abonnement skal kunne kjøre.}

\item{reshID}{Bestillers resh-id. Benyttes for abonnement, ikke utsendinger}

\item{enhetsNivaa}{Enhetsnivå: "Alle" (= hele landet), 'HF', 'RHF'}

\item{nivaaNavn}{navn på det aktuelle HF/RHF, evt. "Alle" (= hele landet). Benyttes for utsendinger.}

\item{Rpakke}{hvilken R-pakke fila som lager rapporten ligger i}
}
\value{
Full path of file produced
}
\description{
Funksjon som produserer rapporten som skal sendes til mottager.
Nødvendig info til Covid-rapporten er reshID og enhetsNivaa
}
