% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIRtabellerBeredskapK2n.R
\name{antallTidInneliggendeBeredskap}
\alias{antallTidInneliggendeBeredskap}
\title{Antall inneliggende per tidsenhet  og enhetsnivå. Filtreringer kan også gjøres.
Detaljerinsnivå er styrt av tilgangsnivå}
\usage{
antallTidInneliggendeBeredskap(
  RegData,
  tidsenhet = "dag",
  erMann = 9,
  resp = 9,
  datoFra = 0,
  bekr = 9,
  skjemastatus = 9,
  dodInt = 9,
  valgtRHF = "Alle",
  velgAvd = 0
)
}
\arguments{
\item{RegData}{dataramme med preprossesserte data}

\item{tidsenhet}{'dag' (standard), 'uke', 'maaned'
('RHF', 'HF', 'ShNavn') resultatene skal vises for}

\item{erMann}{kjønn, 0-kvinne, 1-mann}

\item{resp}{Fått respiratorbehandling: 1-ja, 2-nei}

\item{datoFra}{startdato 'yyyy-mm-dd'}

\item{bekr}{1-bekreftet, 0-mistenkt tilfelle}

\item{skjemastatus}{status på registreringa 0-ingen, 1-kladd, 2-ferdigstilt, 4-slettet, 5-returnert}

\item{dodInt}{død på intensiv 0-nei, 1-ja}

\item{velgAvd}{valgt reshID}
}
\value{

}
\description{
Antall inneliggende per tidsenhet  og enhetsnivå. Filtreringer kan også gjøres.
Detaljerinsnivå er styrt av tilgangsnivå
}
