% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIRberedHjelpefunk.R
\name{BeredskIntensivData}
\alias{BeredskIntensivData}
\title{Koble med intensivdata. Ferdigstilte intensivdata. Kan inneholdeberedskapsdata i kladd}
\usage{
BeredskIntensivData()
}
\description{
Koble med intensivdata. Ferdigstilte intensivdata. Kan inneholdeberedskapsdata i kladd
}
