% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIRfigurerBeredskap.R
\name{FigFordelingKjonnsdelt}
\alias{FigFordelingKjonnsdelt}
\title{Aldersfordeling, kjønnsdelt.}
\usage{
FigFordelingKjonnsdelt(
  RegData,
  valgtVar = "Alder",
  valgtRHF = "Alle",
  bekr = 9,
  skjemastatus = 9,
  resp = 9,
  dodInt = 9,
  erMann = 9,
  grvar = "PatientGender",
  outfile = ""
)
}
\arguments{
\item{TabAlder}{Dataramme med nødvendige figurparametre}
}
\value{

}
\description{
Kan gereraliseres til andre variabler og grupperingsvar.
}
