% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunksjonerDatafilerFHI.R
\name{hentBeredDataFHI}
\alias{hentBeredDataFHI}
\title{Henter intensivberdskapsdata og velger variabler for overføring til FHI}
\usage{
hentBeredDataFHI(personIDvar = "PersonIdBC19Hash", raa = 1, aggP = 1)
}
\arguments{
\item{personIDvar}{angi variabel for personid
'PersonIdBC19Hash' - standard, 'PersonId' - alternativ, PasientGUID - overvåkning}

\item{raa}{hente rådata= 1-ja (standard), 0-nei}

\item{aggP}{hente persondata aggregert til smitteforløp? 1-ja (standard), 0-nei}
}
\value{
Beredskapsdata tilrettelagt for FHI
}
\description{
Henter intensivberdskapsdata og velger variabler for overføring til FHI
}
