% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIRberedHjelpefunk.R
\name{kjor_appNIRbered}
\alias{kjor_appNIRbered}
\title{Kjøre intensivberedskaps app}
\usage{
kjor_appNIRbered(browser = FALSE, logAsJson = FALSE)
}
\arguments{
\item{browser}{vis i nettleser}

\item{logAsJson}{logging}
}
\value{
Brukergrensesnittet til NIRberedskap-appen
}
\description{
Kjøre intensivberedskaps app
}
