% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIRberedHjelpefunk.R
\name{lagStagingData}
\alias{lagStagingData}
\title{Lag staging data for intensivberedskap}
\usage{
lagStagingData()
}
\value{
Character vector of staging files, invisibly
}
\description{
This function makes queries and pre-processing of registry data before
storing relevant staging data. Running this function may take a while so use
with care!
}
