% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIRtabellerBeredskap.R
\name{ManglerIntSkjema}
\alias{ManglerIntSkjema}
\title{Avdelingar som enno har ikkje-ferdigstilte NIR-skjema}
\usage{
ManglerIntSkjema(reshID = 0)
}
\arguments{
\item{reshID}{Avdelingas resh-id}
}
\value{

}
\description{
Avdelingar som enno har ikkje-ferdigstilte NIR-skjema
}
