% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIRberedskFigAndeler.R
\name{NIRberedskFigAndeler}
\alias{NIRberedskFigAndeler}
\title{Funksjon som beregner aggregerte verdier (andeler) for ulike variabler/variabelkombinasjoner}
\usage{
NIRberedskFigAndeler(
  RegData,
  valgtVar = "alder",
  erMann = 9,
  datoFra = 0,
  datoTil = 0,
  enhetsNivaa = "RHF",
  valgtEnhet = "Alle",
  enhetsUtvalg = 0,
  outfile = "",
  lagFig = 1,
  ...
)
}
\arguments{
\item{RegData}{En dataramme med alle nødvendige variabler fra registeret}

\item{valgtVar}{Hvilken variabel som skal visualiseres. Se \strong{Details} for oversikt.
' @param outfile Navn på fil figuren skrives til. Standard: '' (Figur skrives
   til systemets standard utdataenhet (som regel skjerm))}

\item{erMann}{kjønn, 0-kvinne, 1-mann}

\item{datoFra}{startdato 'yyyy-mm-dd'}

\item{datoTil}{sluttdato 'yyyy-mm-dd'}

\item{enhetsUtvalg}{Gjør gruppeutvalg med eller uten sammenlikning. Se \strong{Details} for oversikt.}

\item{lagFig}{Angir om figur skal lages eller ikke 0-ikke lag, 1-lag}

\item{reshID}{Parameter følger fra innlogging helseregister.no og angir
hvilken enhet i spesialisthelsetjenesten brukeren tilhører}

\item{preprosess}{Preprosesser data
FALSE: Nei
TRUE: Ja (Standard)}

\item{hentData}{Gjør spørring mot database
0: Nei, RegData gis som input til funksjonen (Standard)
1: Ja}

\item{dodInt}{Levende/død ut fra intensiv.
0: i live,
1: død,
alle: standard (alle andre verdier)}
}
\value{
Søylediagram (fordeling) av valgt variabel. De enkelte verdiene kan også sendes med.
}
\description{
Denne funksjonen beregner AggVerdier (fordeling) av valgt variabel
filtrert på de utvalg som er gjort. Kan trenge funksjonerne:
}
\details{
Funksjonen benytter funksjonene: NIRPreprossesserBeredsk, NIRUtvalgBeredsk

Argumentet \emph{valgtVar} har følgende valgmuligheter:
   \itemize{
    \item alder: Aldersfordeling, 10-årige grupper
    \item inklKrit: Andeler for de 5 inklusjonskriteriene
    \item liggetid: Liggetid
    \item PrimaryReasonAdmitted: Hovedårsak til intensivopphold
    \item respiratortid: Tid tilbrakt i respirator
    \item spesTiltak: Spesielle tiltak
   }
}
