% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIRsqlPreIfluensa.R
\name{NIRsqlInfluensa}
\alias{NIRsqlInfluensa}
\title{Preprosesser data fra Intensivregisteret}
\usage{
NIRsqlInfluensa(datoFra = "2018-01-01", datoTil = Sys.Date(), preprosess = 1)
}
\arguments{
\item{datoFra}{startdato 'yyyy-mm-dd'}

\item{datoTil}{sluttdato 'yyyy-mm-dd'}

\item{preprosess}{0-rådata, 1-preprosesser}
}
\value{
Dataramme med influensadata fra intensivregisteret
}
\description{
Funksjonen henter influensadata og evt. navner om variabler og beregner nye.
}
