% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIRUtvalgBeredsk.R
\name{NIRUtvalgBeredsk}
\alias{NIRUtvalgBeredsk}
\title{Gjøre utvalg i beredskapsskjema}
\usage{
NIRUtvalgBeredsk(
  RegData,
  datoFra = 0,
  datoTil = 0,
  erMann = 9,
  bekr = 9,
  skjemaStatus = 9,
  dodInt = 9,
  reshID = 0,
  velgRHF = 0,
  velgAvd = 0
)
}
\arguments{
\item{RegData}{data, beredskapsskjema}

\item{datoFra}{startdato 'yyyy-mm-dd'}

\item{datoTil}{sluttdato 'yyyy-mm-dd'}

\item{erMann}{kjønn, 0-kvinne, 1-mann}

\item{bekr}{1-bekreftet, 0-mistenkt tilfelle}

\item{skjemaStatus}{status på registreringa 0-ingen, 1-kladd, 2-ferdigstilt, 4-slettet, 5-returnert}

\item{dodInt}{død på intensiv 0-nei, 1-ja}

\item{reshID}{reshID fra innlogging}

\item{velgAvd}{valgt reshID}

\item{enhetsUtvalg}{enhetsutvalg...}
}
\value{

}
\description{
Returnerer filtrert dataramme og utvalgstekst
}
