% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIRtabellerBeredskap.R
\name{oppsumAldKjTabIkkeKlar}
\alias{oppsumAldKjTabIkkeKlar}
\title{Nøkkeltalltabell - ikke klar}
\usage{
oppsumAldKjTabIkkeKlar(
  RegData,
  erMann = 9,
  bekr = 9,
  skjemastatus = 9,
  dodInt = 9,
  reshID = 0,
  valgtRHF = "Alle"
)
}
\arguments{
\item{RegData}{beredskapsskjema}

\item{erMann}{kjønn, 0-kvinne, 1-mann}

\item{bekr}{1-bekreftet, 0-mistenkt tilfelle}

\item{skjemastatus}{status på registreringa 0-ingen, 1-kladd, 2-ferdigstilt, 4-slettet, 5-returnert}

\item{dodInt}{død på intensiv 0-nei, 1-ja}

\item{reshID}{reshID fra innlogging}
}
\value{

}
\description{
Nøkkeltalltabell - ikke klar
}
