% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIRtabellerBeredskap.R
\name{oppsumFerdigeRegTab}
\alias{oppsumFerdigeRegTab}
\title{Liggetider og antall som har vært i ECMO/respirator}
\usage{
oppsumFerdigeRegTab(RegData, valgtRHF = "Alle", erMann = 9, dodInt = 9)
}
\arguments{
\item{RegData}{beredskapsskjema}

\item{erMann}{kjønn, 0-kvinne, 1-mann}

\item{dodInt}{død på intensiv 0-nei, 1-ja}
}
\value{

}
\description{
Liggetider og antall som har vært i ECMO/respirator
}
