% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIRberedHjelpefunk.R
\name{preprosessBeredVar}
\alias{preprosessBeredVar}
\title{Legger til og formaterer tidsvariable og noen flere}
\usage{
preprosessBeredVar(RegData)
}
\arguments{
\item{RegData}{}
}
\value{

}
\description{
Legger til og formaterer tidsvariable og noen flere
}
