% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIRtabellerBeredskap.R
\name{RisikofaktorerTab}
\alias{RisikofaktorerTab}
\title{Funksjoner for å lage tabeller}
\usage{
RisikofaktorerTab(
  RegData,
  tidsenhet = "Totalt",
  datoTil = Sys.Date(),
  reshID = 0,
  erMann = "",
  bekr = 9,
  skjemaStatus = 9,
  dodInt = 9,
  velgRHF = 0,
  velgAvd = 0
)
}
\arguments{
\item{RegData}{data}

\item{tidsenhet}{'Dag', 'Uke' (standard)}

\item{datoTil}{sluttdato}

\item{reshID}{enhetens resh}
}
\value{

}
\description{
Tabell med oversikt over tilstander som medfører økt risiko ved Coronasmitte
}
