% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIRtabellerBeredskap.R
\name{SentralmaalTab}
\alias{SentralmaalTab}
\title{Tabell med sentralmål, min,maks IQR}
\usage{
SentralmaalTab(
  RegData,
  valgtRHF = "Alle",
  datoFra = "2020-01-01",
  datoTil = Sys.Date(),
  bekr = 9,
  erMann = 9,
  resp = 9,
  dodInt = 9
)
}
\arguments{
\item{RegData}{beredskapsskjema}

\item{datoFra}{startdato 'yyyy-mm-dd'}

\item{datoTil}{sluttdato 'yyyy-mm-dd'}

\item{bekr}{1-bekreftet, 0-mistenkt tilfelle}

\item{erMann}{kjønn, 0-kvinne, 1-mann}

\item{resp}{Fått respiratorbehandling: 1-ja, 2-nei}

\item{dodInt}{død på intensiv 0-nei, 1-ja}
}
\value{

}
\description{
Tabell med sentralmål, min,maks IQR
}
