% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appNIRberedskap.R
\name{server_NIRbered}
\alias{server_NIRbered}
\title{Server logic for the rapRegTemplate app}
\usage{
server_NIRbered(input, output, session)
}
\arguments{
\item{input}{shiny input object}

\item{output}{shiny output object}

\item{session}{shiny session object}
}
\value{
A shiny app server object
}
\description{
Server logic for the rapRegTemplate app
}
