% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIRtabellerBeredskap.R
\name{TabAlder}
\alias{TabAlder}
\title{Aldersfordeling, tabell}
\usage{
TabAlder(RegData, valgtRHF = "Alle")
}
\arguments{
\item{RegData}{datatabell, beredskapsdatt}
}
\value{

}
\description{
Aldersfordeling, tabell
}
\examples{
TabAlder(RegData=CoroData, enhetsNivaa='HF')
}
