% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIRtabellerBeredskap.R
\name{TabTidEnhet}
\alias{TabTidEnhet}
\title{Title}
\usage{
TabTidEnhet(
  RegData,
  tidsenhet = "dag",
  enhetsNivaa = "RHF",
  erMann = 9,
  bekr = 9,
  skjemaStatus = 9,
  dodInt = 9,
  velgRHF = 0,
  velgAvd = 0
)
}
\arguments{
\item{RegData}{dataramme med preprossesserte data}

\item{tidsenhet}{'dag' (standard), 'uke', 'maaned'}

\item{enhetsNivaa}{'RHF', 'HF', 'ShNavn'}
}
\value{

}
\description{
Title
}
