% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIRberedHjelpefunk.R
\name{tilretteleggBeredVar}
\alias{tilretteleggBeredVar}
\title{Legger til og formaterer tidsvariable og noen flere}
\usage{
tilretteleggBeredVar(RegData)
}
\arguments{
\item{RegData}{}
}
\value{

}
\description{
Legger til og formaterer tidsvariable og noen flere
}
