% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIRberedHjelpefunk.R
\name{abonnementBeredsk}
\alias{abonnementBeredsk}
\title{Funksjon som produserer rapporten som skal sendes til mottager.
(The actual call to this function is made through do.call and
has the effect of providing the parameters as class
\emph{list}. Verdier gis inn som listeparametre}
\usage{
abonnementBeredsk(
  rnwFil,
  brukernavn = "beredskap",
  valgtRHF = "Alle",
  reshID = 0,
  enhetsNivaa = "Alle",
  Rpakke = "intensivberedskap"
)
}
\arguments{
\item{rnwFil}{Navn på fila som skal kjøres. Angis MED filending (\emph{dvs "filnavn.Rnw"})}

\item{reshID}{Aktuell reshid}

\item{enhetsNivaa}{Enhetsnivå, 'Alle'-hele landet, 'RHF', 'HF'}

\item{Rpakke}{hvilken R-pakke fila som lager rapporten ligger i}

\item{datoFra}{dato}

\item{parametre}{Liste med valgfrie parametre, avhengig av type rapport}
}
\value{
Full path of file produced
}
\description{
Funksjon som produserer rapporten som skal sendes til mottager.
(The actual call to this function is made through do.call and
has the effect of providing the parameters as class
\emph{list}. Verdier gis inn som listeparametre
}
