% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIRberedHjelpefunk.R
\name{BeredskIntensivData}
\alias{BeredskIntensivData}
\title{Koble med intensivdata. Ferdigstilte intensivdata. Kan inneholde beredskapsdata i kladd}
\usage{
BeredskIntensivData()
}
\value{
Koblet datasett intensiv- og beredskapsskjema
}
\description{
Koble med intensivdata. Ferdigstilte intensivdata. Kan inneholde beredskapsdata i kladd
}
