% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIRfigurerBeredskap.R
\name{FigFordelingKjonnsdelt}
\alias{FigFordelingKjonnsdelt}
\title{Aldersfordeling, kjønnsdelt.}
\usage{
FigFordelingKjonnsdelt(
  RegData,
  valgtVar = "Alder",
  datoFra = "2020-03-01",
  datoTil = Sys.Date(),
  valgtRHF = "Alle",
  bekr = 9,
  skjemastatus = 9,
  resp = 9,
  minN = 0,
  dodInt = 9,
  erMann = 9,
  grvar = "PatientGender",
  outfile = ""
)
}
\arguments{
\item{minN}{Maskerer verdier under <3}

\item{TabAlder}{Dataramme med nødvendige figurparametre}
}
\value{
alders og kjønnsfordelnig
}
\description{
Kan gereraliseres til andre variabler og grupperingsvar.
}
