% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIRberedHjelpefunk.R
\name{henteSamlerapporterBered}
\alias{henteSamlerapporterBered}
\title{Funksjon som produserer rapporten som skal sendes til mottager.}
\usage{
henteSamlerapporterBered(filnavn, rnwFil, reshID = 0, enhetsNivaa = "Alle")
}
\arguments{
\item{filnavn}{dummy}

\item{rnwFil}{Navn på fila som skal kjøres. Angis uten ending (\emph{dvs uten  ".Rnw"})}

\item{reshID}{Aktuell reshid}

\item{enhetsNivaa}{Enhetsnivå, 'Alle'-hele landet, 'RHF', 'HF'}
}
\value{
Full path of file produced
}
\description{
Funksjon som produserer rapporten som skal sendes til mottager.
}
