% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIRberedHjelpefunk.R
\name{henteSamlerapporterBered}
\alias{henteSamlerapporterBered}
\title{Funksjon som produserer rapporten som skal sendes til mottager.}
\usage{
henteSamlerapporterBered(
  filnavn,
  rnwFil,
  reshID = 0,
  Rpakke = "intensivberedskap",
  datoFra = Sys.Date() - 180,
  datoTil = Sys.Date()
)
}
\arguments{
\item{filnavn}{dummy}

\item{rnwFil}{Navn på fila som skal kjøres. Angis uten ending (\emph{dvs uten  ".Rnw"})}

\item{reshID}{Aktuell reshid}

\item{Rpakke}{hvilken R-pakke fila som lager rapporten ligger i}

\item{datoFra}{dato}

\item{parametre}{Liste med valgfrie parametre, avhengig av type rapport}
}
\value{
Full path of file produced
}
\description{
Funksjon som produserer rapporten som skal sendes til mottager.
}
