% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIRtabellerInfluensa.R
\name{InfluensaUkeRHF}
\alias{InfluensaUkeRHF}
\title{Influensatilfeller per uke, regionvis og hele landet}
\usage{
InfluensaUkeRHF(
  RegData,
  bekr = 9,
  skjemastatus = 9,
  dodInt = 9,
  erMann = 9,
  sesong = "2021-22",
  alleUker = 1
)
}
\arguments{
\item{RegData}{Influensadata}

\item{bekr}{1:bekreftet 0: mistenkt tilfelle}

\item{sesong}{influensasesong}

\item{alleUker}{0: viser uker med hendelser, 1: viser alle uker}

\item{ferdigstilt}{ferdigstilt registrering}
}
\description{
Influensatilfeller per uke, regionvis og hele landet
}
