% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InfluensadataFHI.R
\name{lagInfluDataFHI}
\alias{lagInfluDataFHI}
\title{Henter data og tilrettelegger filer for overføring til FHI}
\usage{
lagInfluDataFHI()
}
\value{

}
\description{
Henter data og tilrettelegger filer for overføring til FHI
}
