% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIRtabellerBeredskap.R
\name{ManglerIntSkjema}
\alias{ManglerIntSkjema}
\title{Avdelingar som enno har ikkje-ferdigstilte NIR-skjema for ferdigstilte beredskapsskjema}
\usage{
ManglerIntSkjema(reshID = 0, datoFra = "2020-03-01", datoTil = Sys.Date())
}
\arguments{
\item{reshID}{Avdelingas resh-id}
}
\value{

}
\description{
Avdelingar som enno har ikkje-ferdigstilte NIR-skjema for ferdigstilte beredskapsskjema
}
