% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIRberedskDataSQL.R
\name{NIRberedskDataSQL}
\alias{NIRberedskDataSQL}
\title{Henter data fra pårørendeskjema registrert for Intensiv og kobler til
noen variabler fra hovedskjema.}
\usage{
NIRberedskDataSQL()
}
\arguments{
\item{medH}{kobler på variabler fra hovedskjema}
}
\value{
Henter dataramma RegData for Intensivregisteret
}
\description{
Henter data for Intensivregisterets database
}
