% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIRberedskDataSQL.R
\name{NIRberedskDataSQL}
\alias{NIRberedskDataSQL}
\title{Hente data fra beredskapsskjema}
\usage{
NIRberedskDataSQL(datoFra = "2020-03-01", datoTil = Sys.Date(), kobleInt = 0)
}
\arguments{
\item{datoFra}{fra og med, innleggelsesdato}

\item{datoTil}{til og med, innleggelsesdato}

\item{kobleInt}{koble på data fra intensivskjema 0 - nei (standard), 1 - ja}
}
\value{
Henter dataramme
}
\description{
Henter data fra beredskapsskjema registrert for Intensiv og kobler til
noen variabler fra hovedskjema.
}
