% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIRPreprossesserBeredsk.R
\name{NIRPreprosessBeredsk}
\alias{NIRPreprosessBeredsk}
\title{Preprosesser data fra Intensivregisteret}
\usage{
NIRPreprosessBeredsk(RegData = RegData, kobleInt = 0, skjema = "bered")
}
\arguments{
\item{RegData}{Beredskapsskjema}

\item{kobleInt}{koble på data fra intensivskjema. Hvis koblede data, filtreres registreringer uten intensivskjema bort.}

\item{skjema}{hvilket skjema data som skal preprosesseres tilhører
'bered': beredskap (korona), 'influ': influensa}
}
\value{
Data En liste med det filtrerte datasettet (og sykehusnavnet som tilsvarer reshID, ikke pt)
}
\description{
Denne funksjonen navner om variabler og beregner evt. nye.
Funksjonen lager også et offentlig datasett som kan benyttes til beregning
av kvalitetsindikatorer og som kan legges ved pakken
}
