% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIRtabellerBeredskap.R
\name{oppsumLiggetiderTab}
\alias{oppsumLiggetiderTab}
\title{Antall som er eller har vært i ECMO/respirator}
\usage{
oppsumLiggetiderTab(RegData, valgtRHF = "Alle")
}
\arguments{
\item{RegData}{beredskapsskjema}
}
\value{

}
\description{
Antall som er eller har vært i ECMO/respirator
}
