% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIRtabellerBeredskap.R
\name{statusECMOrespTab}
\alias{statusECMOrespTab}
\title{Antall som er  i ECMO/respirator}
\usage{
statusECMOrespTab(RegData, valgtRHF = "Alle", erMann = 9, bekr = 9)
}
\arguments{
\item{RegData}{beredskapsskjema}
}
\description{
Antall som er  i ECMO/respirator
}
