% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIRtabellerBeredskap.R
\name{statusECMOrespTab}
\alias{statusECMOrespTab}
\title{Antall som er eller har vært i ECMO/respirator}
\usage{
statusECMOrespTab(RegData)
}
\arguments{
\item{RegData}{beredskapsskjema}
}
\value{

}
\description{
Antall som er eller har vært i ECMO/respirator
}
