% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIRtabellerBeredskap.R
\name{tabRegForsinkelse}
\alias{tabRegForsinkelse}
\title{Tabell som viser fordeling av registreringsforsinkelse per enhet}
\usage{
tabRegForsinkelse(
  RegData,
  innUt = 1,
  datoFra = "2020-03-01",
  datoTil = Sys.Date(),
  pst = 1
)
}
\arguments{
\item{RegData}{registerdata, IKKE personaggregert}

\item{innUt}{1-forsinkelse innreg, 2-forsinkelse, ferdigstille utskr.}

\item{datoFra}{dato, fra og med}

\item{datoTil}{dato, til og med}

\item{pst}{0-antall, 1-prosent}
}
\description{
Tabell som viser fordeling av registreringsforsinkelse per enhet
}
