% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIRtabellerBeredskap.R
\name{TabTidEnhet}
\alias{TabTidEnhet}
\title{Antall tilfeller for valgt tidsenhet og enhetsnivå. Filtreringer kan også gjøres.}
\usage{
TabTidEnhet(
  RegData,
  tidsenhet = "dag",
  erMann = 9,
  resp = 9,
  datoFra = 0,
  bekr = 9,
  skjemastatus = 9,
  dodInt = 9,
  valgtRHF = "Alle",
  velgAvd = 0
)
}
\arguments{
\item{RegData}{dataramme med preprossesserte data}

\item{tidsenhet}{'dag' (standard), 'uke', 'maaned'}

\item{erMann}{kjønn, 0-kvinne, 1-mann}

\item{resp}{Fått respiratorbehandling: 1-ja, 2-nei}

\item{datoFra}{Vis hendelser fra og med dato}

\item{bekr}{1-bekreftet, 0-mistenkt tilfelle}

\item{skjemastatus}{status på registreringa 0-ingen, 1-kladd, 2-ferdigstilt, 4-slettet, 5-returnert}

\item{dodInt}{død på intensiv 0-nei, 1-ja}

\item{velgAvd}{valgt reshID}

\item{enhetsNivaa}{'RHF', 'HF', 'ShNavn'}
}
\value{
Tabell med antall for hver tidsenhet og enhet
}
\description{
Antall tilfeller for valgt tidsenhet og enhetsnivå. Filtreringer kan også gjøres.
}
