% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIRtabellerBeredskap.R
\name{TabTidEnhet}
\alias{TabTidEnhet}
\title{Title}
\usage{
TabTidEnhet(RegData, tidsenhet = "dag", enhetsNivaa = "RHF")
}
\arguments{
\item{RegData}{dataramme med preprossesserte data}

\item{tidsenhet}{'dag' (standard), 'uke', 'maaned'}

\item{enhetsNivaa}{'RHF', 'HF', 'ShNavn'}
}
\value{

}
\description{
Title
}
